mesh clean up and modularization
